add steps to research workflow
line 61:  not clear how we could search for related URLS, and  how to report back if you found a related URL in the pipeline so as not to duplicate work.   This is something that is not uncommon. Whether this search for duplicates should happen earlier in the process was something that we also discussed.